+++ client/src/main/scala/denimcoat/svg/TextFacade.scala
@@ -12,13 +12,13 @@ class TextFacade(val svg: SVG, val element: Text) extends StylableElementFacade[
 
   def text_=(text: String): Unit = element.textContent = text
 
-  def x: Int = Try(element.getAttribute("x").toInt).fold(Function.const(0), identity)
+  def x: Double = Try(element.getAttribute("x").toDouble).fold(Function.const(0.0), identity)
 
-  def x_=(x: Int): Unit = element.setAttribute("x", x.toString)
+  def x_=(x: Double): Unit = element.setAttribute("x", x.toString)
 
-  def y: Int = Try(element.getAttribute("y").toInt).fold(Function.const(0), identity)
+  def y: Double = Try(element.getAttribute("y").toDouble).fold(Function.const(0.0), identity)
 
-  def y_=(y: Int): Unit = element.setAttribute("y", y.toString)
+  def y_=(y: Double): Unit = element.setAttribute("y", y.toString)
 
   def fill: String = element.getAttribute("fill")
 
@@ -32,7 +32,7 @@ object TextFacade {
   def getById(svg: SVG, id: String): TextFacade =
     new TextFacade(svg, SvgUtils.getSvgElementById[Text](id))
 
-  def create(svg: SVG, id: String, x: Int, y: Int): TextFacade = {
+  def create(svg: SVG, id: String, x: Double, y: Double): TextFacade = {
     val text = new TextFacade(svg, SvgUtils.createSvgElement[Text]("text"))
     text.id = id
     text.x = x
